requirements.txt
fastapi
uvicorn
python-multipart
pillow
cryptography